feat: 테마 추가/삭제 관리자용 분리
