+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -1,0 +1,39 @@
+package roomescape.controller;
+
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.ThemeRequest;
+import roomescape.dto.ThemeResponse;
+import roomescape.service.ThemeService;
+
+@RestController
+@RequestMapping("/admin/themes")
+public class AdminThemeController {
+    private static final String LOCATION_DEFAULT_VALUE = "admin/themes/";
+
+    private final ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ThemeResponse> addTheme(@RequestBody ThemeRequest request) {
+        ThemeResponse response = themeService.addTheme(request);
+        return ResponseEntity.created(URI.create(LOCATION_DEFAULT_VALUE + response.id()))
+                .body(response);
+    }
+
+    @DeleteMapping("/{themeId}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable("themeId") Long themeId) {
+        themeService.delete(themeId);
+        return ResponseEntity.noContent()
+                .build();
+    }
+}
