+++ src/main/java/roomescape/controller/ThemeController.java
@@ -1,25 +1,18 @@
 package roomescape.controller;
 
-import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
 import roomescape.service.ThemeService;
 
 @RestController
 @RequestMapping("/themes")
 public class ThemeController {
-    private static final String LOCATION_DEFAULT_VALUE = "/themes/";
 
     private final ThemeService themeService;
 
@@ -27,23 +20,9 @@ public class ThemeController {
         this.themeService = themeService;
     }
 
-    @PostMapping
-    public ResponseEntity<ThemeResponse> addTheme(@RequestBody ThemeRequest request) {
-        ThemeResponse response = themeService.addTheme(request);
-        return ResponseEntity.created(URI.create(LOCATION_DEFAULT_VALUE + response.id()))
-                .body(response);
-    }
-
     @GetMapping
     public ResponseEntity<List<ThemeResponse>> getPopularThemes(@RequestParam("endDate") LocalDate endDate) {
         List<ThemeResponse> responses = themeService.getPopularThemes(endDate);
         return ResponseEntity.ok(responses);
-    }
-
-    @DeleteMapping("/{themeId}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable("themeId") Long themeId) {
-        themeService.delete(themeId);
-        return ResponseEntity.noContent()
-                .build();
     }
 }
