+++ src/test/java/roomescape/controller/ThemeControllerTest.java
@@ -1,44 +1,29 @@
 package roomescape.controller;
 
+import static org.hamcrest.Matchers.is;
+
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import java.util.HashMap;
-import java.util.Map;
+import java.time.LocalDate;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.jdbc.Sql;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class ThemeControllerTest {
 
     @Test
-    void 테마_추가() {
-        Map<String, String> themeParams = new HashMap<>();
-        themeParams.put("name", "방탈출1");
-        themeParams.put("description", "다함께 탈출해요 방탈출.");
-        themeParams.put("thumbnail", "https://asdfsdf.sdfs");
+    @Sql(scripts = {"/data.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
+    void 인기_테마를_조회한다() {
+        LocalDate endDate = LocalDate.of(2026, 5, 5);
 
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
-                .body(themeParams)
-                .when().post("/themes")
+                .when().get("/themes?endDate=" + endDate)
                 .then().log().all()
-                .statusCode(201);
-    }
-
-    @Test
-    void 테마_삭제() {
-        Map<String, String> themeParams = new HashMap<>();
-        themeParams.put("name", "방탈출11");
-        themeParams.put("description", "다함께 탈출해요 방탈출.");
-        themeParams.put("thumbnail", "https://asdfsdf.sdfs");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(themeParams)
-                .when().post("/themes")
-                .then().log().all()
-                .statusCode(201);
+                .statusCode(200)
+                .body("[0].name", is("공포의 저택"));
     }
 }
